Updates for gbfs v3.0 release
