+++ gbfs-java-model/src/test/java/org/entur/gbfs/v3_0/manifest/ManifestTest.java
@@ -1,0 +1,12 @@
+package org.entur.gbfs.v3_0.manifest;
+
+import org.entur.gbfs.TestBase;
+import org.entur.gbfs.v3_0.manifest.GBFSManifest;
+import org.junit.jupiter.api.Test;
+
+class ManifestTest extends TestBase {
+    @Test
+    void testUnmarshal() {
+        assertUnmarshalDoesNotThrow("v3_0/manifest.json", GBFSManifest.class);
+    }
+}
