--- gbfs-java-model/src/test/java/org/entur/gbfs/v3_0_RC/gbfs_versions/GBFSVersionsTest.java
@@ -1,11 +1,0 @@
-package org.entur.gbfs.v3_0_RC.gbfs_versions;
-
-import org.entur.gbfs.TestBase;
-import org.junit.jupiter.api.Test;
-
-class GBFSVersionsTest extends TestBase {
-    @Test
-    void testUnmarshal() {
-        assertUnmarshalDoesNotThrow("v3_0_RC/gbfs_versions.json", GBFSGbfsVersions.class);
-    }
-}
