--- gbfs-java-model/src/test/java/org/entur/gbfs/v3_0_RC/station_information/StationInformationTest.java
@@ -1,11 +1,0 @@
-package org.entur.gbfs.v3_0_RC.station_information;
-
-import org.entur.gbfs.TestBase;
-import org.junit.jupiter.api.Test;
-
-class StationInformationTest extends TestBase {
-    @Test
-    void testUnmarshal() {
-        assertUnmarshalDoesNotThrow("v3_0_RC/station_information.json", GBFSStationInformation.class);
-    }
-}
