+++ gbfs-java-model/src/test/resources/v3_0/manifest.json
@@ -1,0 +1,35 @@
+{
+  "last_updated":1667004473,
+  "ttl":0,
+  "version":"3.0",
+  "data":{
+    "datasets":[
+      {
+        "system_id":"example_berlin",
+        "versions":[
+          {
+            "version":"2.0",
+            "url":"https://berlin.example.com/gbfs/2/gbfs"
+          },
+          {
+            "version":"3.0",
+            "url":"https://berlin.example.com/gbfs/3/gbfs"
+          }
+        ]
+      },
+      {
+        "system_id":"example_paris",
+        "versions":[
+          {
+            "version":"2.0",
+            "url":"https://paris.example.com/gbfs/2/gbfs"
+          },
+          {
+            "version":"3.0",
+            "url":"https://paris.example.com/gbfs/3/gbfs"
+          }
+        ]
+      }
+    ]
+  }
+}
