--- gbfs-java-model/src/test/resources/v3_0_RC/station_information.json
@@ -1,89 +1,0 @@
-{
-  "last_updated": 1640887163,
-  "ttl": 0,
-  "version": "3.0-RC",
-  "data": {
-    "stations": [
-      {
-        "station_id": "pga",
-        "name": [
-          {
-            "text": "Parking garage A",
-            "language": "en"
-          }
-        ],
-        "lat": 12.345678,
-        "lon": 45.678901,
-        "station_opening_hours": "Su-Th 05:00-22:00; Fr-Sa 05:00-01:00",
-        "parking_type": "underground_parking",
-        "parking_hoop": false,
-        "contact_phone": "+33109874321",
-        "is_charging_station": true,
-        "vehicle_type_dock_capacity": [
-          {
-            "vehicle_type_id": "abc123",
-            "count": 7
-          },
-          {
-            "vehicle_type_id": "def456",
-            "count": 0
-          }
-        ]
-      },
-      {
-        "station_id": "station12",
-        "name": [
-          {
-            "text": "SE Belmont & SE 10th",
-            "language": "en"
-          }
-        ],
-        "lat": 45.516445,
-        "lon": -122.655775,
-        "is_valet_station": false,
-        "is_virtual_station": true,
-        "is_charging_station": false,
-        "station_area": {
-          "type": "MultiPolygon",
-          "coordinates": [
-            [
-              [
-                [
-                  -122.655775,
-                  45.516445
-                ],
-                [
-                  -122.655705,
-                  45.516445
-                ],
-                [
-                  -122.655705,
-                  45.516495
-                ],
-                [
-                  -122.655775,
-                  45.516495
-                ],
-                [
-                  -122.655775,
-                  45.516445
-                ]
-              ]
-            ]
-          ]
-        },
-        "capacity": 16,
-        "vehicle_type_area_capacity": [
-          {
-            "vehicle_type_id": "abc123",
-            "count": 7
-          },
-          {
-            "vehicle_type_id": "def456",
-            "count": 8
-          }
-        ]
-      }
-    ]
-  }
-}
