--- gbfs-java-model/src/test/resources/v3_0_RC/station_status.json
@@ -1,71 +1,0 @@
-{
-  "last_updated": 1640887163,
-  "ttl": 0,
-  "version": "3.0-RC",
-  "data": {
-    "stations": [
-      {
-        "station_id": "station1",
-        "is_installed": true,
-        "is_renting": true,
-        "is_returning": true,
-        "last_reported": 1609866125,
-        "num_docks_available": 3,
-        "num_docks_disabled" : 1,
-        "vehicle_docks_available": [
-          {
-            "vehicle_type_ids": [ "abc123", "def456" ],
-            "count": 2
-          },
-          {
-            "vehicle_type_ids": [ "def456" ],
-            "count": 1
-          }
-        ],
-        "num_vehicles_available": 1,
-        "num_vehicles_disabled": 2,
-        "vehicle_types_available": [
-          {
-            "vehicle_type_id": "abc123",
-            "count": 1
-          },
-          {
-            "vehicle_type_id": "def456",
-            "count": 0
-          }
-        ]
-      },
-      {
-        "station_id": "station2",
-        "is_installed": true,
-        "is_renting": true,
-        "is_returning": true,
-        "last_reported": 1609866106,
-        "num_docks_available": 8,
-        "num_docks_disabled" : 1,
-        "vehicle_docks_available": [
-          {
-            "vehicle_type_ids": [ "abc123" ],
-            "count": 6
-          },
-          {
-            "vehicle_type_ids": [ "def456" ],
-            "count": 2
-          }
-        ],
-        "num_vehicles_available": 6,
-        "num_vehicles_disabled": 1,
-        "vehicle_types_available": [
-          {
-            "vehicle_type_id": "abc123",
-            "count": 2
-          },
-          {
-            "vehicle_type_id": "def456",
-            "count": 4
-          }
-        ]
-      }
-    ]
-  }
-}
