--- gbfs-java-model/src/test/resources/v3_0_RC/system_alerts.json
@@ -1,43 +1,0 @@
-{
-  "last_updated": 1604519393,
-  "ttl": 60,
-  "version": "3.0-RC",
-  "data": {
-    "alerts": [
-      {
-        "alert_id": "21",
-        "type": "station_closure",
-        "station_ids": [
-          "123",
-          "456",
-          "789"
-        ],
-        "times": [
-          {
-            "start": 1604448000,
-            "end": 1604674800
-          }
-        ],
-        "url": [
-          {
-            "text": "https://example.com/more-info",
-            "language": "en"
-          }
-        ],
-        "summary": [
-          {
-            "text": "Disruption of Service",
-            "language": "en"
-          }
-        ],
-        "description": [
-          {
-            "text": "The three stations on Broadway will be out of service from 12:00am Nov 3 to 3:00pm Nov 6th to accommodate road work",
-            "language": "en"
-          }
-        ],
-        "last_updated": 1604198100
-      }
-    ]
-  }
-}
