--- gbfs-java-model/src/test/resources/v3_0_RC/system_pricing_plans.json
@@ -1,78 +1,0 @@
-{
-  "last_updated": 1640887163,
-  "ttl": 0,
-  "version": "3.0-RC",
-  "data": {
-    "plans": [
-      {
-        "plan_id": "plan2",
-        "name": [
-          {
-            "text": "One-Way",
-            "language": "en"
-          }
-        ],
-        "currency": "USD",
-        "price": 2.00,
-        "reservation_price_per_min": 0.15,
-        "is_taxable": false,
-        "description": [
-          {
-            "text": "Includes 10km, overage fees apply after 10km.",
-            "language": "en"
-          }
-        ],
-        "per_km_pricing": [
-          {
-            "start": 10,
-            "rate": 1.00,
-            "interval": 1,
-            "end": 25
-          },
-          {
-            "start": 25,
-            "rate": 0.50,
-            "interval": 1
-          },
-          {
-            "start": 25,
-            "rate": 3.00,
-            "interval": 5
-          }
-        ]
-      },
-      {
-        "plan_id": "plan3",
-        "name": [
-          {
-            "text": "Simple Rate",
-            "language": "en"
-          }
-        ],
-        "currency": "CAD",
-        "price": 3.00,
-        "is_taxable": true,
-        "description": [
-          {
-            "text": "$3 unlock fee, $0.25 per kilometer and 0.50 per minute.",
-            "language": "en"
-          }
-        ],
-        "per_km_pricing": [
-          {
-            "start": 0,
-            "rate": 0.25,
-            "interval": 1
-          }
-        ],
-        "per_min_pricing": [
-          {
-            "start": 0,
-            "rate": 0.50,
-            "interval": 1
-          }
-        ]
-      }
-    ]
-  }
-}
