--- gbfs-java-model/src/test/resources/v3_0_RC/vehicle_status.json
@@ -1,63 +1,0 @@
-{
-  "last_updated":1640887163,
-  "ttl":0,
-  "version":"3.0-RC",
-  "data":{
-    "vehicles":[
-      {
-        "vehicle_id":"973a5c94-c288-4a2b-afa6-de8aeb6ae2e5",
-        "last_reported":1609866109,
-        "lat":12.345678,
-        "lon":56.789012,
-        "is_reserved":false,
-        "is_disabled":false,
-        "vehicle_type_id":"abc123",
-        "rental_uris": {
-          "android": "https://www.example.com/app?vehicle_id=973a5c94-c288-4a2b-afa6-de8aeb6ae2e5&platform=android&",
-          "ios": "https://www.example.com/app?vehicle_id=973a5c94-c288-4a2b-afa6-de8aeb6ae2e5&platform=ios"
-        }
-      },
-      {
-        "vehicle_id":"987fd100-b822-4347-86a4-b3eef8ca8b53",
-        "last_reported":1609866204,
-        "is_reserved":false,
-        "is_disabled":false,
-        "vehicle_type_id":"def456",
-        "current_range_meters":6543.0,
-        "station_id":"86",
-        "pricing_plan_id":"plan3"
-      },
-      {
-        "vehicle_id":"45bd3fb7-a2d5-4def-9de1-c645844ba962",
-        "last_reported":1609866109,
-        "lat":12.345678,
-        "lon":56.789012,
-        "is_reserved":false,
-        "is_disabled":false,
-        "vehicle_type_id":"abc123",
-        "current_range_meters":400000.0,
-        "available_until":"2021-05-17T15:00:00Z",
-        "home_station_id":"station1",
-        "vehicle_equipment":[
-          "child_seat_a",
-          "winter_tires"
-        ]
-      },
-      {
-        "vehicle_id":"d4521def-7922-4e46-8e1d-8ac397239bd0",
-        "last_reported":1609866204,
-        "is_reserved":false,
-        "is_disabled":false,
-        "vehicle_type_id":"def456",
-        "current_fuel_percent":0.7,
-        "current_range_meters":6543.0,
-        "station_id":"86",
-        "pricing_plan_id":"plan3",
-        "home_station_id":"146",
-        "vehicle_equipment":[
-          "child_seat_a"
-        ]
-      }
-    ]
-  }
-}
